Correction answer in Exercise 1.3
(I don't know why I wrote 3...)
There are 5 predicates: 
woman
man
person 
loves
father

In http://www.learnprolognow.org/lpnpage.php?pagetype=html&pageid=lpn-htmlse1 
KB2 is:
happy(yolanda). 
   listens2Music(mia). 
   listens2Music(yolanda):-  happy(yolanda). 
   playsAirGuitar(mia):-  listens2Music(mia). 
   playsAirGuitar(yolanda):-  listens2Music(yolanda).

and later in the text we can see the next lines:  Another way of looking at KB2 is to say that it consists of three predicates (or procedures). The three predicates are:

   listens2Music 
   happy 
   playsAirGuitar

So, I understand that a predicate is the "name" of the fact or rule no matters how many times appears this